# internal/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nops API",
    "version": "1",
    "license": { "name": "Apache 2.0", "identifier": "Apache-2.0" },
    "description": "Read jobs and deployments, act on them, and manage ACL policies, binding rules and tokens. A token does what its ACL policies allow, with the dashboard's rules."
  },
  "servers": [{ "url": ".." }],
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/api/deployments/{id}": {
      "get": {
        "summary": "One deployment with its redacted diff, hook runs, and events",
        "operationId": "getDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "responses": {
          "200": {
            "description": "The deployment.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentDetail" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "A token from the dashboard's Administration page." }
    },
    "parameters": {
...
      "DeploymentID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "No valid token: it is missing, unknown, expired, or revoked.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The token is valid, and its ACL policies do not allow this.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The job, deployment, ACL policy or token does not exist.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ServerError": {
        "description": "Nops hit an error and logged it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "Deployment": {
        "type": "object",
        "required": ["id", "namespace", "job", "state", "policy", "spec_hash", "commit_sha", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "namespace": { "type": "string" },
          "job": { "type": "string" },
          "state": { "type": "string" },
          "policy": { "type": "string", "enum": ["auto", "approval", "none"] },
          "spec_hash": { "type": "string" },
          "commit_sha": { "type": "string" },
          "commit_subject": { "type": "string" },
          "commit_author": { "type": "string" },
          "error": { "type": "string" },
          "decided_by": { "type": "string" },
          "decided_at": { "type": "string", "format": "date-time" },
          "retry_of": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Event": {
        "type": "object",
        "required": ["time", "from", "to", "actor"],
        "properties": {
          "time": { "type": "string", "format": "date-time" },
          "from": { "type": "string" },
          "to": { "type": "string" },
          "actor": { "type": "string" },
          "accessor_id": { "type": "string", "description": "The token the action was made with. Absent for Nops itself." },
          "message": { "type": "string" }
        }
      },
      "HookRun": {
        "type": "object",
        "required": ["phase", "job", "state"],
        "properties": {
          "phase": { "type": "string" },
          "job": { "type": "string" },
          "state": { "type": "string" },
          "error": { "type": "string" },
          "started_at": { "type": "string", "format": "date-time" },
          "finished_at": { "type": "string", "format": "date-time" }
        }
      },
      "DeploymentDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Deployment" },
          {
            "type": "object",
            "required": ["events", "hook_runs"],
            "properties": {
              "diff": { "type": "object", "description": "The redacted plan diff, as Nomad's JSON. Absent when there is none." },
              "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
              "hook_runs": { "type": "array", "items": { "$ref": "#/components/schemas/HookRun" } }
            }
          }
        ]
      },
...
    }
  }
}